Improved and expanded testing
Streaming appears to work locally now

Improved testing using byte array

Improved testing using a local geotiff

improved testing for s3StreamReading

Updated and expanded testing

Made some corrections to reading and streaming bytes

Removed some debugging from TiffTagsReader

continuing work on s3reader

renamed test

+++ s3/src/test/scala/geotrellis/spark/io/s3/MockS3Client.scala
@@ -103,7 +103,7 @@ class MockS3Client() extends S3Client with LazyLogging {
     }
   }
   
-  def readRange(start: Int, end: Int, getObjectRequest: GetObjectRequest): S3Object = ???
+  def readRange(start: Long, end: Long, getObjectRequest: GetObjectRequest): S3Object = ???
 
   def putObject(r: PutObjectRequest): PutObjectResult = this.synchronized {
     logger.debug(s"PUT ${r.getKey}")
